Add POST Request to accept text from client and log to server
Issue: I am using body-parser to parse the request coming in and access
the text.  When I am logging the request from my server it is displaying
as undefined.  When I am displaying the entire request object, the body
object is empty.  Any reason this could be the case? @chinedufn