Use a pool of clients to connect to DB INSERT is working
Problem before was that I was connecting to the wrong DB therefore it
was not finding my table.  Now I added process.env.PGUSER to contain the
user, and PGDATABASE to hold the name of the database to connect to.

The Current issue is that I would like to have the text being sent to
the server through the POST request use the INSERT to place the text and
the timestamp into the Database.  Currently the way I have it client is
not defined.  Do I need to move the pool.connect() function and have it
wrao the handling of the post request? @chinedufn