React-Router working properly and create two input fields for login page
I had to move the code found in App.js to index.js in order for
React-Router to work, can now link to '/login'.  That page has two input
text boxes for email and password.

The Next step is to add a submit button, pass the email and password to
the server to then store it in the DB.