Add Button to Login Page to send login info to Server
Problem: I am trying to add a second route to handle the post request
from the Login page with '/#/login'. When there are two routes active
and the new one for login is placed before the route for '/', it still
fails to recognize the new one for Login and executes the route for '/'.
When I remove the code for the '/' route, I receive this message in the
console: "POST http://localhost:3000/ 404 Not Found" and the '/#/login'
route is not executed.  Any idea what could be happening here ?
@chinedufn