Attempt to Modularize React Components Issue with 'signUp' function
In line 4 of submit-button.js I am creating a function that is supposed
to take in a url and then open a XHR and sent a POST request to my
server.  The problem I am having is creating passing in a url to my
function in the file that uses Submit Button.  In SignUp.js i try to
create the url(on line 4) and pass it in to the  signUp function as a
parameter in order to reduce code and make this button reusable
@chinedufn anything I am missing here?